% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band_pass_filter_app.R
\name{band_pass_filter_app}
\alias{band_pass_filter_app}
\title{Band-pass Filter Shiny App}
\usage{
band_pass_filter_app(launch.browser = FALSE)
}
\arguments{
\item{launch.browser}{Logical. If TRUE, the app will automatically open in the default web browser. Defaults to FALSE.}
}
\value{
A Shiny app interface for applying a frequency filter.
}
\description{
This function launches a Shiny app that allows users to visualize either the mean spectrum or the spectrogram before applying a frequency filter to a Wave object.
}
\examples{
\dontrun{
# Launch the app
band_pass_filter_app()

# Example usage after launching the app:
# 1. Select a wave object from the R environment.
# 2. Specify the the limits of the band-pass filter (high-pass, low-pass) in kHz, as needed.
# 3. Apply the filter and save the new wave object.
}
}
