% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorcode_oscillogram_app.R
\name{colorcode_oscillogram_app}
\alias{colorcode_oscillogram_app}
\title{Color-Coded Oscillogram Shiny App}
\usage{
colorcode_oscillogram_app(launch.browser = FALSE)
}
\arguments{
\item{launch.browser}{Logical. If TRUE, the app will automatically open in the default web browser. Defaults to FALSE.}
}
\value{
A Shiny app for visualizing, customizing, and downloading color-coded oscillograms.
}
\description{
This Shiny app allows users to visualize an oscillogram and paint selected sections with different colors using a color-blind-safe palette. The user can also customize the display options such as adding a scale bar, showing axes, and downloading the plot as a PNG file.
}
\examples{
if (interactive()) {
  colorcode_oscillogram_app()
}
}
