% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_app.R
\name{downsample_app}
\alias{downsample_app}
\title{Downsample Audio Shiny App}
\usage{
downsample_app(launch.browser = FALSE)
}
\arguments{
\item{launch.browser}{Logical. If TRUE, the app will automatically open in the default web browser. Defaults to FALSE.}
}
\value{
A Shiny app interface for downsampling and visualizing audio wave objects.
}
\description{
This function launches a Shiny app that allows users to downsample audio (wave objects) and visualize the mean spectrum. Users can select an existing wave object from the R environment, specify a new maximum frequency (kHz), downsample the audio, and save the downsampled wave object with a new name.
}
\examples{
\dontrun{
downsample_app()
}
}
