% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_wave_app.R
\name{import_wave_app}
\alias{import_wave_app}
\title{Interactive Audio Import Shiny App}
\usage{
import_wave_app(launch.browser = FALSE)
}
\arguments{
\item{launch.browser}{Logical. If TRUE, the app will automatically open in the default web browser. Defaults to FALSE.}
}
\value{
This function launches a Shiny app for importing audio files. The processed wave object is saved to the global environment with the name provided by the user.
}
\description{
This function launches a Shiny app that allows users to import and process audio files (WAV, MP3, or WAC format) into the R environment. The app supports stereo audio by allowing the user to select a channel (left or right) for processing, and it automatically removes any offset in the wave object. Users can assign a custom name to the processed wave object, which is then saved in the global environment.
}
\examples{
\dontrun{
# Launch the app
import_wave_app()
}
}
