% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanspectrum_plotly.R
\name{meanspectrum_plotly}
\alias{meanspectrum_plotly}
\title{Interactive Mean Spectrum}
\usage{
meanspectrum_plotly(
  wave,
  background = "#274C77",
  foreground = "white",
  hover_bgcolor = "white",
  hover_fontcolor = "black"
)
}
\arguments{
\item{wave}{A Wave object.}

\item{background}{Background color.}

\item{foreground}{Color for plot lines and spectral signatures.}

\item{hover_bgcolor}{Color for the hover background.}

\item{hover_fontcolor}{Color for the hover text.}
}
\value{
An interactive Mean Spectrum
}
\description{
Interactive Mean Spectrum
}
\examples{
\dontrun{
data(coryphoda)
meanspectrum_plotly(coryphoda)
}
}
