% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_waves_app.R
\name{merge_waves_app}
\alias{merge_waves_app}
\title{Shiny App to Merge Wave Objects}
\usage{
merge_waves_app()
}
\value{
A Shiny app for merging `Wave` objects
}
\description{
This function launches a Shiny app that allows users to select multiple `Wave` objects from the R environment, merge them over time,
and save the merged object back to R memory or download it as a `.wav` file.
}
\examples{
if (interactive()) {
  merge_waves_app()
}
}
