% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_meanspectra_static_app.R
\name{multi_meanspectra_static_app}
\alias{multi_meanspectra_static_app}
\title{Multi-Power Spectra Static Shiny App}
\usage{
multi_meanspectra_static_app(launch.browser = FALSE)
}
\arguments{
\item{launch.browser}{Logical. If TRUE, the app will automatically open in the default web browser. Defaults to FALSE.}
}
\value{
A Shiny app for visualizing wave oscillograms and mean spectra.
}
\description{
This Shiny app allows users to create multiple power spectra from selections in the oscillogram.
The app provides options to download individual or combined visualizations.
}
\examples{
\dontrun{
if (interactive()) {
  multi_meanspectra_static_app()
}
}
}
