% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multioscillo_app.R
\name{multioscillo_app}
\alias{multioscillo_app}
\title{Multi-Oscillogram Shiny App}
\usage{
multioscillo_app(launch.browser = FALSE)
}
\arguments{
\item{launch.browser}{Logical. If TRUE, the app will automatically open in the default web browser. Defaults to FALSE.}
}
\value{
A Shiny app interface for plotting and saving multi-oscillograms of selected wave objects.
}
\description{
This function launches a Shiny app that allows users to select multiple wave objects, plot their oscillograms, and save the resulting plots as a PNG file. Users can control the maximum duration of the waves to be plotted, set the scale bar, and specify image dimensions.
}
\examples{
\dontrun{
multioscillo_app()
}
}
