% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot_app.R
\name{multiplot_app}
\alias{multiplot_app}
\title{Multiplot (Spectrogram, Oscillogram, and Mean Power Spectrum) Shiny App}
\usage{
multiplot_app(launch.browser = FALSE)
}
\arguments{
\item{launch.browser}{Logical. If TRUE, the app will automatically open in the default web browser. Defaults to FALSE.}
}
\value{
A Shiny app for visualizing, customizing, and downloading spectrogram and oscillogram plots.
}
\description{
This Shiny app allows users to visualize a combined plot of a spectrogram, mean spectrum, and oscillogram for a selected audio file (Wave object). Users can adjust settings such as noise cutoff, image dimensions, and background transparency, and can download the resulting plot as a PNG file.
}
\examples{
\dontrun{
if (interactive()) {
  multiplot_app()
}
}
}
