% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot_ggplot.R
\name{multiplot_ggplot}
\alias{multiplot_ggplot}
\title{Multiplot including Spectrogram, Mean Power Spectrum, and}
\usage{
multiplot_ggplot(
  wave,
  cutoff = -35,
  overlap = 95,
  zeropad = 200,
  scale.type = "dB",
  heights = c(7, 3),
  oscillo.label = ""
)
}
\arguments{
\item{wave}{A `Wave` object. The sound data to be plotted.}

\item{cutoff}{Numeric. Amplitude cutoff in dB below the peak. Values below the cutoff will not be visible in the spectrogram. Default is -35 dB.}

\item{overlap}{Numeric. Window overlap percentage for the spectrogram. Default is 95.}

\item{zeropad}{Numeric. Zero padding for FFT in the spectrogram. Default is 200.}

\item{scale.type}{Character. Scale for the amplitude in the spectrogram, either "dB" or "Linear". Default is "dB".}

\item{heights}{Numeric vector. The relative heights of the spectrogram/mean spectrum plot and the oscillogram. Default is `c(7,3)`.}

\item{oscillo.label}{Character. A title for the Y-axis of the oscillogram. Default is an empty string.}
}
\value{
A plot with aligned spectrogram, mean spectrum, and oscillogram.
}
\description{
Multiplot including Spectrogram, Mean Power Spectrum, and
}
\examples{
\dontrun{
data(coryphoda)
multiplot_ggplot(coryphoda)
}
}
