% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillogram_ggplot.R
\name{oscillogram_ggplot}
\alias{oscillogram_ggplot}
\title{Oscillogram with ggplot2}
\usage{
oscillogram_ggplot(wave, y.title = "Relative Amplitude", x.title = "")
}
\arguments{
\item{wave}{A Wave object}

\item{y.title}{Character. A title for the Y-axis.}

\item{x.title}{Character. A title for the X-axis.}
}
\value{
An oscillogram plot.
}
\description{
Oscillogram with ggplot2
}
\examples{
\dontrun{
oscillgoram_ggplot(coryphoda)
}
}
