% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillogram_interactive.R
\name{oscillogram_interactive}
\alias{oscillogram_interactive}
\title{Interactive Oscillogram}
\usage{
oscillogram_interactive(wave, title = "", line_color = "black")
}
\arguments{
\item{wave}{A wave object containing the audio data to be plotted.}

\item{title}{A string specifying the title of the plot. Default is an empty string.}

\item{line_color}{A string specifying the color of the waveform line. Default is 'black'.}
}
\value{
A Plotly object representing the interactive oscillogram.
}
\description{
This function generates an interactive oscillogram plot for a given wave object using Plotly.
The plot displays the waveform over time with customizable options for the title and line color.
}
\examples{
\dontrun{
oscillogram_interactive(coryphoda_wave, title = "Coryphoda albidicollis")
}
}
