% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillogram_zoom_app.R
\name{oscillogram_zoom_app}
\alias{oscillogram_zoom_app}
\title{Oscillogram + Zoomed Stack Shiny App}
\usage{
oscillogram_zoom_app(dark.mode = TRUE)
}
\arguments{
\item{dark.mode}{Logical, if TRUE the app will use a dark theme.}
}
\value{
A Shiny app that displays an oscillogram and zoomed sections below, with options to save the plot as a PNG.
}
\description{
This Shiny app provides an interface to visualize an oscillogram along with zoomed-in portions of the waveform stacked below.
The user can select portions of the oscillogram to zoom in on, display either axes or a scale bar, and save the resulting image.
}
\examples{
\dontrun{
multioscillo_app()
}
if (interactive()) {
  oscillogram_zoom_app(dark.mode = TRUE)
}
}
