% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_stats_app.R
\name{spectral_stats_app}
\alias{spectral_stats_app}
\title{Spectral Statistics Shiny App}
\usage{
spectral_stats_app(launch.browser = FALSE)
}
\arguments{
\item{launch.browser}{Logical. If TRUE, the app will automatically open in the default web browser. Defaults to FALSE.}
}
\value{
A Shiny app for analyzing and visualizing spectral statistics of audio wave objects.
}
\description{
This Shiny app allows users to analyze spectral statistics of audio data from wave objects. Users can select a wave object, input additional metadata (such as Specimen ID and sound type), and configure parameters for spectral analysis, including amplitude threshold and bandwidth. The app generates a Plotly plot of the spectrum, showing key spectral features such as peak frequency, bandwidth, and -20 dB threshold. Users can download the results in HTML or CSV format and save the data frame to the R environment.
}
\examples{
 \dontrun{
if (interactive()) {
  spec_stats_app()
}
}
}
