% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram_plotly.R
\name{spectrogram_plotly}
\alias{spectrogram_plotly}
\title{Interactive Spectrogram with Standardized Resolution}
\usage{
spectrogram_plotly(
  wave,
  floor = -50,
  background = "#274C77",
  foreground = "white",
  hover_bgcolor = "white",
  hover_fontcolor = "black",
  overlap = 50,
  zero_padding = 0
)
}
\arguments{
\item{wave}{A Wave object.}

\item{floor}{Background noise level, in dB FS.}

\item{background}{Background color.}

\item{foreground}{Color for plot lines and spectral signatures.}

\item{hover_bgcolor}{Color for the hover background.}

\item{hover_fontcolor}{Color for the hover text.}

\item{overlap}{FFT overlap.}

\item{zero_padding}{Number of zeroes to be added around the signal to improve frequency resolution.}
}
\value{
An interactive spectrogram.
}
\description{
Interactive Spectrogram with Standardized Resolution
}
\examples{
\dontrun{
data(coryphoda)
spectrogram_plotly(coryphoda)
}
}
