% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_plot_app.R
\name{spectrum_plot_app}
\alias{spectrum_plot_app}
\title{Spectrum Plot Shiny App}
\usage{
spectrum_plot_app(launch.browser = FALSE)
}
\arguments{
\item{launch.browser}{Logical. If TRUE, the app will automatically open in the default web browser. Defaults to FALSE.}
}
\value{
A Shiny app that generates and visualizes the spectrum plot of an audio wave object.
}
\description{
This Shiny app allows users to visualize a spectrum plot for a selected audio wave object. It includes customizable options such as window length, overlap, window function, and amplitude scale (linear, dB, dBFS). Users can also adjust the axis positions, choose the number of frequency ticks, and add plot parameters or summary statistics to the plot.
}
\examples{
 \dontrun{
if (interactive()) {
  spectrum_plot_app()
}
}
}
