% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_stats_hq_app.R
\name{temporal_stats_hq_app}
\alias{temporal_stats_hq_app}
\title{Temporal Statistics HQ Shiny App}
\usage{
temporal_stats_hq_app(launch.browser = FALSE)
}
\arguments{
\item{launch.browser}{Logical. If TRUE, the app will automatically open in the default web browser. Defaults to FALSE.}
}
\value{
A Shiny app that analyzes temporal statistics of acoustic waveforms.
}
\description{
This Shiny app performs temporal statistics analysis on wave objects with "HQ" sounds (i.e., near pure tones such as those made by most crickets). It allows the user to configure multiple parameters such as smoothing window, overlap, and detection thresholds to analyze the temporal patterns of acoustic signals. The app outputs interactive visualizations of the results and provides data tables summarizing trains, motifs, and summary statistics.
}
\examples{
 \dontrun{
if (interactive()) {
  temp_stats_hiq_app()
}
}
}
