% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_stats_lq.R
\name{temporal_stats_lq}
\alias{temporal_stats_lq}
\title{Temporal Statistics Function for Acoustic Analysis of LQ signals}
\usage{
temporal_stats_lq(
  wave,
  specimen_id = "",
  ssmooth = 100,
  peakfinder_ws = 50,
  peakfinder_threshold = 0.005,
  max_train_gap = 0.5,
  max_peak_gap = 0.01,
  detection_threshold = 0.01,
  norm_env = TRUE
)
}
\arguments{
\item{wave}{A `Wave` object from the `tuneR` package representing the acoustic signal.}

\item{specimen_id}{A character string representing the identifier for the specimen (or recording) being analyzed.}

\item{ssmooth}{Numeric value representing the smoothing window size (in samples) for the amplitude envelope.}

\item{peakfinder_ws}{Numeric value specifying the window size (in samples) for the peak detection algorithm.}

\item{peakfinder_threshold}{Numeric value specifying the amplitude threshold for peak detection, as a proportion of the maximum amplitude.}

\item{max_train_gap}{Numeric value representing the maximum allowed gap (in seconds) between peaks to consider them part of the same train.}

\item{max_peak_gap}{Numeric value representing the maximum allowed gap (in seconds) between peaks to consider them as part of the same group.}

\item{detection_threshold}{Numeric value specifying the minimum amplitude for a peak to be considered valid.}

\item{norm_env}{Logical value indicating whether to normalize the amplitude envelope between 0 and 1. Defaults to `TRUE`.}
}
\value{
A list containing the following:
\item{plot}{An interactive plotly plot showing the waveform envelope, detected peaks, trains, and motifs.}
\item{peak_data}{A tibble containing detailed information about detected peaks, including their times and periods.}
\item{train_data}{A tibble containing details about detected trains, including start and end times, durations, peak rates, and gaps.}
\item{motif_data}{A tibble summarizing the detected motifs, including motif duration, train counts, and duty cycle.}
\item{params}{A tibble of input parameters for reference.}
}
\description{
This function calculates temporal statistics from a wave object,
including peak detection, train detection, and motif grouping.
It is designed for analyzing LQ acoustic signals,
detecting peaks in amplitude envelopes,
and measuring the duration and timing of motifs and trains in the waveform.
}
\details{
New metrics provided by this function include the Temporal Complexity Index (within motif),
Temporal Excursion (within train), and Dynamic Excursion (within train).
}
\examples{
\dontrun{
data(coryphoda)
# result <- temporal_stats_lq(coryphoda)
# plot(result$plot)
}

}
